pyproject.toml
[tool.poetry]
name = "dalle3-img-gen"
version = "0.1.0"
description = ""
authors = ["Rinku <shark4rkcoding@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
openai = "^1.12.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
